- Attempts to resolve DURACLOUD-94 (http://fedora-commons.org/jira/browse/DURACLOUD-94) by ensuring that streams are closed and files get deleted - Adds a minor update to the image conversion test to allow passing in a credential, note that code must be updated prior to running tests (which are not part of the normal build cycle anyway because an install of ImageMagick is required) to include password. This is done rather than trying to include the unit-test-db (which is not an OSGi bundle) into the OSGi container just for this test.
git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@593 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ services/imageconversionservice/src/main/java/org/duracloud/services/image/ConversionManager.java
@@ -179,23 +179,30 @@ public class ConversionManager {
         Content sourceContent =
             contentStore.getContent(sourceSpaceId, contentId);
         InputStream sourceStream = sourceContent.getStream();
-        File sourceFile = writeSourceToFile(sourceStream, contentId);
 
-        // Perform conversion
-        File convertedFile = convertImage(sourceFile);
-
+        File sourceFile = null;
+        File convertedFile = null;
         try {
+            sourceFile = writeSourceToFile(sourceStream, contentId);
+
+            // Perform conversion
+            convertedFile = convertImage(sourceFile);
+
             // Store the converted file in the destination space
             storeConvertedContent(convertedFile,
                                   sourceContent.getMetadata());
         } finally {
             // Delete source and converted files from work directory
-            if (!sourceFile.delete()) {
-                sourceFile.deleteOnExit();
+            if(sourceFile != null) {
+                if (!sourceFile.delete()) {
+                    sourceFile.deleteOnExit();
+                }
             }
 
-            if (!convertedFile.delete()) {
-                convertedFile.deleteOnExit();
+            if(convertedFile != null) {
+                if (!convertedFile.delete()) {
+                    convertedFile.deleteOnExit();
+                }
             }
         }
     }
@@ -255,13 +262,16 @@ public class ConversionManager {
         sourceFile.createNewFile();
         FileOutputStream sourceOut = new FileOutputStream(sourceFile);
 
-        long sizeCopied = IOUtils.copyLarge(sourceStream, sourceOut);
-        if(sizeCopied <= 0) {
-            throw new IOException("Unable to copy any bytes from file " +
-                fileName);
+        try {
+            long sizeCopied = IOUtils.copyLarge(sourceStream, sourceOut);
+            if(sizeCopied <= 0) {
+                throw new IOException("Unable to copy any bytes from file " +
+                    fileName);
+            }
+        } finally {
+            sourceStream.close();
+            sourceOut.close();
         }
-
-        sourceOut.close();
         return sourceFile;
     }
